requirements-dev.txt
pytest==8.3.4
PyYAML==6.0.2
jsonschema==4.23.0
